refactor to one shared disclosure with different classes
