+++ src/views/tutorial-view/components/variant-dropdown-disclosure/types.ts
@@ -1,21 +1,6 @@
-import { ReactNode } from 'react'
-import {
-	TutorialVariant,
-	TutorialVariantOption,
-} from 'views/tutorial-view/utils/variants/types'
+import { TutorialVariant } from 'views/tutorial-view/utils/variants/types'
 
 export interface VariantDropdownDisclosureProps {
 	variant: TutorialVariant
 	className?: string
-}
-
-export interface VariantDropdownDisclosureWithLabelProps {
-	id: string
-	text: string
-	children: ReactNode
-}
-
-export interface VariantDropdownDisclosureItemProps {
-	variant: TutorialVariant
-	activeOption: TutorialVariantOption
 }
